+++ deno_tag.ts
@@ -1,3 +1,5 @@
+// Copyright 2021 Hugo Daniel Henriques Oliveira Gomes. All rights reserved.
+// Licensed under the EUPL
 import { denoTag } from "./code.ts";
 
 /**