+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.1.0] - 2021-03-04
+### Added
+- README.md to the examples folder
+- `defaults.ts` with the default implementation for the bundler
+- Support the new `Deno.emit` compiler API
+- Support for Deno v1.8.0
 - Export the `DenoTagOptions` interface (no longer exported as a type object)
 - Updated CLI usage examples in the README file